Implement exit to all POSIX platforms
