+++ internal/sh/builtin/exit.go
@@ -1,37 +1,55 @@
-//+build !linux
-
 package builtin
 
 import (
 	"fmt"
-	"runtime"
+	"os"
+	"strconv"
 
+	"github.com/NeowayLabs/nash/errors"
 	"github.com/NeowayLabs/nash/sh"
 )
 
 type (
-	exitNotSupportedFn struct {
-		err error
+	exitFn struct {
+		status int
 	}
 )
 
-func newExit() *exitNotSupportedFn {
-	return &exitNotSupportedFn{
-		err: fmt.Errorf("exit is not implemented on OS: %s", runtime.GOOS),
-	}
+func newExit() *exitFn {
+	return &exitFn{}
 }
 
-func (e *exitNotSupportedFn) ArgNames() []string {
-	panic(e.err)
-	return []string{}
+func (e *exitFn) ArgNames() []string {
+	return []string{"status"}
 }
 
-func (e *exitNotSupportedFn) Run() (sh.Obj, error) {
-	panic(e.err)
+func (e *exitFn) Run() (sh.Obj, error) {
+	os.Exit(e.status)
 	return nil, nil //Unrecheable code
 }
 
-func (e *exitNotSupportedFn) SetArgs(args []sh.Obj) error {
-	panic(e.err)
+func (e *exitFn) SetArgs(args []sh.Obj) error {
+	if len(args) != 1 {
+		return errors.NewError("exit expects one argument")
+	}
+
+	obj := args[0]
+	if obj.Type() != sh.StringType {
+		return errors.NewError(
+			"exit expects a status string, but a %s was provided",
+			obj.Type(),
+		)
+	}
+	statusstr := obj.(*sh.StrObj).Str()
+	status, err := strconv.Atoi(statusstr)
+	if err != nil {
+		return fmt.Errorf(
+			"exit:linux:error[%s] converting status[%s] to int",
+			err,
+			statusstr,
+		)
+
+	}
+	e.status = status
 	return nil
 }
